[change] Hash calling_station_id
